refactor(detail): extraer hooks de dialogs y emails
